and more refactor
